whiskerguard/regulatory/requirements.txt
pytest==8.3.5
requests==2.32.3
allure-pytest==2.14.0
locust==2.24.0
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-rerunfailures==12.0
pytest-timeout==2.2.0
pytest-assume==2.4.3
pytest-ordering==0.6
pytest-dependency==0.5.1
pytest-sugar==1.0.0
pytest-metadata==3.0.0
pytest-base-url==2.0.0
pytest-variables==3.0.0
pytest-env==1.1.3
pytest-randomly==3.15.0
pytest-cov==4.1.0 